profile/requirements.txt
vllm
numpy
pyyaml